pyproject.toml
[project]
name = "superstore-use"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.7.9",
    "flask>=3.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "modal>=1.3.0.post1",
    "playwright>=1.57.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]
